requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0.post1
SQLAlchemy==2.0.23
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
httpx==0.25.1
Jinja2==3.1.2
aiofiles==23.2.1
typer==0.9.0
pytest==7.4.4
